Moved keyboard shortcuts to Toolbox
